rename `certificates` package to `certificate`
