+++ cmd/certificates/auth.go
@@ -18,7 +18,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/zeebo/errs"
 
-	"storj.io/storj/pkg/certificates/authorization"
+	"storj.io/storj/pkg/certificate/authorization"
 	"storj.io/storj/pkg/process"
 )
