+++ cmd/certificates/claims.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/zeebo/errs"
 
-	"storj.io/storj/pkg/certificates/authorization"
+	"storj.io/storj/pkg/certificate/authorization"
 	"storj.io/storj/pkg/process"
 )
