+++ cmd/identity/main.go
@@ -16,7 +16,7 @@ import (
 
 	"storj.io/storj/internal/fpath"
 	"storj.io/storj/internal/version"
-	"storj.io/storj/pkg/certificates/certificatesclient"
+	"storj.io/storj/pkg/certificate/certificatesclient"
 	"storj.io/storj/pkg/cfgstruct"
 	"storj.io/storj/pkg/identity"
 	"storj.io/storj/pkg/peertls/extensions"
